Implementing missing sorts, adding chart and table displaying of every sort

+++ Sort/src/main/java/sort/model/sorts/Sort.java
@@ -6,6 +6,6 @@ package sort.model.sorts;
  */
 public interface Sort {
     
-    public int runSort(int[] toSort);
+    public long runSort(int[] toSort);
     
 }
